Remove Unique Identifier service layer and move all functionality down into UniqueIdentifierDao
