+++ src/main/java/uk/gov/ea/datareturns/domain/model/validation/auditors/controlledlist/UniqueIdentifierAuditor.java
@@ -1,7 +1,7 @@
 package uk.gov.ea.datareturns.domain.model.validation.auditors.controlledlist;
 
 import org.springframework.stereotype.Component;
-import uk.gov.ea.datareturns.domain.jpa.service.UniqueIdentifierService;
+import uk.gov.ea.datareturns.domain.jpa.dao.UniqueIdentifierDao;
 import uk.gov.ea.datareturns.domain.model.validation.constraints.controlledlist.ControlledListAuditor;
 
 import javax.inject.Inject;
@@ -15,7 +15,8 @@ import java.util.Objects;
 @Component
 public class UniqueIdentifierAuditor implements ControlledListAuditor {
     @Inject
-    UniqueIdentifierService dao;
+    UniqueIdentifierDao dao;
+
     /* (non-Javadoc)
      * @see uk.gov.ea.datareturns.domain.model.validation.constraints.controlledlist.ControlledListAuditor#isValid(java.lang.Object)
      */
